refactor(audit-log): inline embed builders into AuditLogManager
Move buildCommandExecuteEmbed and buildSettingsChangeEmbed from the
standalone auditLogEmbeds.ts utility into private methods on
AuditLogManager, co-locating embed logic with the class that owns it.
Inline CommandExecutePayload, SettingsChangePayload, and AuditLogAction
types; delete auditLogEmbeds.ts and remove its barrel re-export.

Also simplify create-mute.ts to use writeSettings instead of the
writeSettingsTransaction + trx pattern.
